Add files via upload
This repository contains a detailed case study on feature engineering for Delhivery, one of India's leading logistics and supply chain companies. The objective of this project is to demonstrate the process of creating, refining, and selecting features that improve the performance of machine learning models for logistics and delivery optimization.